Optimize userinstall.sh by keeping it more BASH
The simple `while read` loops allow for reading input (be it a file, -
the output from command/process substitution, or something else) one
line at a time. This approach is ideal for smaller amounts of data like
the output of lspci(8) and lsusb(8).

I'd have done the same for the first grep(1) call which looks for '.'
but I wasn't sure why you were use grep(1) just to look for _something_
so I left it alone. If you're just looking for the result of `acpi`
simply _having_ a value, a better tool for the job, which is already in
BASH, is the `-n` test flag, which returns 0 for a non-empty value.

```bash
if [ -n "$(acpi)" ]; then
    # config file to indicate being a laptop
    echo "device appears to be a laptop"
    iconf -i islaptop 1
fi
```

There's actually a tool called laptop-detect(1) which you may have
available in Arch Linux; I recommend looking into that as an
alternative to `acpi`; it's probably lighter-weight, at the very least.
It tends to be available out-of-the-box in Debian- and Ubuntu-based
distributions, but I'm not sure about Arch ones.